Struct layout_thread_2020::LayoutThread

pub struct LayoutThread {

    pub(crate) id: PipelineId,
    pub(crate) url: ServoUrl,
    pub(crate) stylist: Stylist,
    pub(crate) is_iframe: bool,
    pub(crate) constellation_chan: IpcSender<LayoutMsg>,
    pub(crate) script_chan: IpcSender<ConstellationControlMsg>,
    pub(crate) time_profiler_chan: ProfilerChan,
    pub(crate) image_cache: Arc<dyn ImageCache>,
    pub(crate) font_context: Arc<FontContext>,
    pub(crate) first_reflow: Cell<bool>,
    pub(crate) generation: Cell<u32>,
    pub(crate) box_tree: RefCell<Option<Arc<BoxTree>>>,
    pub(crate) fragment_tree: RefCell<Option<Arc<FragmentTree>>>,
    pub(crate) epoch: Cell<Epoch>,
    pub(crate) viewport_size: Size2D<Au>,
    pub(crate) scroll_offsets: RefCell<HashMap<ExternalScrollId, Vector2D<f32, LayoutPixel>>>,
    pub(crate) webrender_image_cache: Arc<RwLock<FnvHashMap<(ServoUrl, UsePlaceholder), WebRenderImageInfo>>>,
    pub(crate) registered_painters: RegisteredPaintersImpl,
    pub(crate) compositor_api: CrossProcessCompositorApi,
    pub(crate) paint_time_metrics: PaintTimeMetrics,
    pub(crate) last_iframe_sizes: RefCell<FnvHashMap<BrowsingContextId, Size2D<f32, CSSPixel>>>,
    pub(crate) debug: DebugOptions,

}

Information needed by layout.

Fields
id: PipelineId

The ID of the pipeline that we belong to.

url: ServoUrl

The URL of the pipeline that we belong to.

stylist: Stylist

Performs CSS selector matching and style resolution.

is_iframe: bool

Is the current reflow of an iframe, as opposed to a root window?

constellation_chan: IpcSender<LayoutMsg>

The channel on which messages can be sent to the constellation.

script_chan: IpcSender<ConstellationControlMsg>

The channel on which messages can be sent to the script thread.

time_profiler_chan: ProfilerChan

The channel on which messages can be sent to the time profiler.

image_cache: Arc<dyn ImageCache>

Reference to the script thread image cache.

font_context: Arc<FontContext>

A FontContext to be used during layout.

first_reflow: Cell<bool>

Is this the first reflow in this LayoutThread?

generation: Cell<u32>

Starts at zero, and increased by one every time a layout completes. This can be used to easily check for invalid stale data.

box_tree: RefCell<Option<Arc<BoxTree>>>

The box tree.

fragment_tree: RefCell<Option<Arc<FragmentTree>>>

The fragment tree.

epoch: Cell<Epoch>

A counter for epoch messages

viewport_size: Size2D<Au>

The size of the viewport. This may be different from the size of the screen due to viewport constraints.

scroll_offsets: RefCell<HashMap<ExternalScrollId, Vector2D<f32, LayoutPixel>>>

Scroll offsets of nodes that scroll.

webrender_image_cache: Arc<RwLock<FnvHashMap<(ServoUrl, UsePlaceholder), WebRenderImageInfo>>>registered_painters: RegisteredPaintersImpl

The executors for paint worklets.

compositor_api: CrossProcessCompositorApi

Cross-process access to the Compositor API.

paint_time_metrics: PaintTimeMetrics

Paint time metrics.

last_iframe_sizes: RefCell<FnvHashMap<BrowsingContextId, Size2D<f32, CSSPixel>>>

The sizes of all iframes encountered during the last layout operation.

debug: DebugOptions

Debug options, copied from configuration to this LayoutThread in order to avoid having to constantly access the thread-safe global options.

Implementations

impl LayoutThread

pub(crate) fn new( id: PipelineId, url: ServoUrl, is_iframe: bool, constellation_chan: IpcSender<ConstellationMsg>, script_chan: IpcSender<ConstellationControlMsg>, image_cache: Arc<dyn ImageCache>, resource_threads: ResourceThreads, system_font_service: Arc<SystemFontServiceProxy>, time_profiler_chan: ProfilerChan, compositor_api: CrossProcessCompositorApi, paint_time_metrics: PaintTimeMetrics, window_size: WindowSizeData, ) -> LayoutThread

pub(crate) fn build_shared_style_context<'a>( &'a self, guards: StylesheetGuards<'a>, snapshot_map: &'a SnapshotMap, animation_timeline_value: f64, animations: &DocumentAnimationSet, traversal_flags: TraversalFlags, ) -> SharedStyleContext<'a>

pub(crate) fn build_layout_context<'a>( &'a self, guards: StylesheetGuards<'a>, snapshot_map: &'a SnapshotMap, origin: ImmutableOrigin, animation_timeline_value: f64, animations: &DocumentAnimationSet, stylesheets_changed: bool, use_rayon: bool, ) -> LayoutContext<'a>

pub(crate) fn load_all_web_fonts_from_stylesheet_with_guard( &self, stylesheet: &DocumentStyleSheet, guard: &SharedRwLockReadGuard<'_>, )

pub(crate) fn handle_reflow(&mut self, data: &mut ScriptReflowResult)

The high-level routine that performs layout.

pub(crate) fn update_scroll_node_state(&self, state: &ScrollState)

pub(crate) fn perform_post_style_recalc_layout_passes( &self, fragment_tree: Arc<FragmentTree>, reflow_goal: &ReflowGoal, document: Option<&ServoLayoutDocument<'_>>, context: &mut LayoutContext<'_>, )

pub(crate) fn profiler_metadata(&self) -> Option<TimerMetadata>

Returns profiling information which is passed to the time profiler.

pub(crate) fn cancel_animations_for_nodes_not_in_fragment_tree( animations: &DocumentAnimationSet, root: &FragmentTree, )

Cancel animations for any nodes which have been removed from fragment tree. TODO(mrobinson): We should look into a way of doing this during flow tree construction. This also doesn’t yet handles nodes that have been reparented.

pub(crate) fn update_iframe_sizes( &self, new_iframe_sizes: FnvHashMap<BrowsingContextId, Size2D<f32, CSSPixel>>, )

Update the recorded iframe sizes of the contents of layout and when these sizes changes, send a message to the constellation informing it of the new sizes.

pub(crate) fn handle_viewport_change( &mut self, window_size_data: WindowSizeData, guards: &StylesheetGuards<'_>, ) -> bool

Update layout given a new viewport. Returns true if the viewport changed or false if it didn’t.